+++ src/components/Sidebar/SidebarPanel/WeaponList/WeaponList.vue
@@ -1,32 +1,50 @@
 <script setup lang="ts">
 import type { IWeapon } from '@/data/interfaces';
+import { computed } from 'vue';
 import WeaponListEntry from './WeaponListEntry.vue';
 
 const props = defineProps<{
     weapons: IWeapon[],
+    limitWeapons: boolean,
 }>();
 
 const emit = defineEmits<{
     (e: "entryClicked", weapon: IWeapon): void,
+    (e: "showAllWeapons"): void,
 }>();
+
+const limitedWeapons = computed(() => {
+    const weapons = props.weapons;
+    return props.limitWeapons && weapons.length > 100 ? weapons.slice(0, 100) : weapons;
+});
 
 function onEntryClicked(weapon: IWeapon) {
     emit("entryClicked", weapon);
 }
+
+function onShowAllClick() {
+    emit("showAllWeapons");
+}
 </script>
 
 <template>
-    <div class="list">
+    <nav class="list">
         <WeaponListEntry
-            v-for="weapon of props.weapons"
+            v-for="weapon of limitedWeapons"
             :key="weapon.hash"
             :weapon="weapon"
             @entry-clicked="onEntryClicked"
         ></WeaponListEntry>
-    </div>
+        <footer class="limited" v-if="props.limitWeapons">
+            <i class="text">Results are currently limited.</i>
+            <button class="show" @click="onShowAllClick">
+                <i>Show All</i>
+            </button>
+        </footer>
+    </nav>
 </template>
 
-<style scoped>
+<style scoped lang="less">
 .list {
     overflow-x: hidden;
     overflow-y: scroll;
@@ -34,5 +52,35 @@ function onEntryClicked(weapon: IWeapon) {
     flex-direction: column;
     gap: 16px;
     padding: 16px;
+}
+
+.results-limited-mixin() {
+    color: #fafafa;
+    font-size: 12px;
+    font-weight: bold;
+}
+
+.limited {
+    align-self: center;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 8px;
+
+    opacity: 0.7;
+}
+
+.text {
+    .results-limited-mixin();
+}
+
+.show {
+    cursor: pointer;
+    border: none;
+    box-shadow: none;
+    background-color: transparent;
+
+    text-decoration: underline;
+    .results-limited-mixin();
 }
 </style>